break up giant steps file
