+++ pilates/workflows/steps/shared.py
@@ -1464,1280 +1464,3 @@ def _execute_atlas_postprocess(
     _warn_missing_coupler_inputs(coupler, raw_outputs, context)
     return run_postprocessor(postprocessor, raw_outputs, workspace)
 
-
-def make_urbansim_preprocess_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the UrbanSim preprocess step function.
-
-    This step prepares land-use inputs by materializing the UrbanSim mutable
-    data directory, ensuring warm-start activities (if enabled), and making
-    required input tables available for the land-use runner.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing preprocess outputs.
-
-    Returns
-    -------
-    callable
-        Step function for UrbanSim preprocess.
-    """
-
-    def _log_inputs(
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> Dict[str, Any]:
-        if state.is_start_year() and settings.activitysim.warm_start_activities:
-            logger.info("[Main] Running warm start activities for ActivitySim.")
-            warm_start_activities(settings, state, workspace)
-        return {}
-
-    def _log_outputs(
-        outputs: UrbanSimPreprocessOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        for short_name, path, description in outputs._iter_record_items():
-            log_and_set_output(
-                key=short_name,
-                path=str(path),
-                description=description,
-                coupler=coupler,
-                **_urbansim_output_facet_meta(
-                    short_name, forecast_year=state.forecast_year
-                ),
-            )
-        usim_data_dir = outputs.usim_mutable_data_dir
-        usim_input_fname = settings.urbansim.input_file_template.format(
-            region_id=settings.urbansim.region_mappings["region_to_region_id"][
-                settings.run.region
-            ]
-        )
-        usim_input_path = usim_data_dir / usim_input_fname
-        if usim_input_path.exists():
-            log_and_set_output(
-                key=USIM_DATASTORE_BASE_H5,
-                path=str(usim_input_path),
-                description="UrbanSim base datastore for preprocessing",
-                coupler=coupler,
-                profile_file_schema=True,
-                h5_container=True,
-                hash_tables="if_unchanged",
-                **_urbansim_output_facet_meta(
-                    USIM_DATASTORE_BASE_H5, forecast_year=state.forecast_year
-                ),
-            )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="urbansim",
-        phase="preprocess",
-        outputs_class=UrbanSimPreprocessOutputs,
-        component_getter=lambda factory, state: factory.get_preprocessor(
-            "urbansim", state, WorkflowState.Stage.land_use
-        ),
-        component_executor=_execute_preprocess,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "urbansim_preprocess", outputs
-        ),
-        input_logger=_log_inputs,
-        output_logger=_log_outputs,
-    )
-
-
-def make_urbansim_run_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the UrbanSim run step function.
-
-    This step executes the UrbanSim land-use simulation for the forecast year
-    and produces the UrbanSim datastore output.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing run outputs.
-
-    Returns
-    -------
-    callable
-        Step function for UrbanSim run.
-    """
-
-    def _log_outputs(
-        outputs: UrbanSimRunOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        if outputs.usim_datastore_h5 is not None:
-            log_and_set_output(
-                key=USIM_DATASTORE_H5,
-                path=str(outputs.usim_datastore_h5),
-                description=(
-                    f"UrbanSim datastore output for year {state.forecast_year}"
-                ),
-                coupler=coupler,
-                profile_file_schema=True,
-                h5_container=True,
-                hash_tables="if_unchanged",
-                **_urbansim_output_facet_meta(
-                    USIM_DATASTORE_H5, forecast_year=state.forecast_year
-                ),
-            )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="urbansim",
-        phase="run",
-        outputs_class=UrbanSimRunOutputs,
-        component_getter=lambda factory, state: factory.get_runner(
-            "urbansim", state, WorkflowState.Stage.land_use
-        ),
-        component_executor=_execute_urbansim_run,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "urbansim_run", outputs
-        ),
-        output_logger=_log_outputs,
-    )
-
-
-def make_urbansim_postprocess_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the UrbanSim postprocess step function.
-
-    This step merges UrbanSim outputs into the input datastore used by
-    downstream models and prepares the HDF5 for the next stage.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing postprocess outputs.
-
-    Returns
-    -------
-    callable
-        Step function for UrbanSim postprocess.
-    """
-
-    def _log_outputs(
-        outputs: UrbanSimPostprocessOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        for short_name, path, description in outputs._iter_record_items():
-            if short_name.startswith(USIM_INPUT_ARCHIVE_PREFIX):
-                log_output_only(
-                    key=short_name,
-                    path=str(path),
-                    description=description,
-                    profile_file_schema=True,
-                    h5_container=True,
-                    hash_tables="if_unchanged",
-                    **_urbansim_output_facet_meta(
-                        short_name, forecast_year=state.forecast_year
-                    ),
-                )
-        if outputs.usim_datastore_h5 is not None:
-            log_and_set_output(
-                key=USIM_DATASTORE_H5,
-                path=str(outputs.usim_datastore_h5),
-                description=(
-                    "UrbanSim datastore prepared for next iteration "
-                    f"(year {state.forecast_year})"
-                ),
-                coupler=coupler,
-                profile_file_schema=True,
-                h5_container=True,
-                hash_tables="if_unchanged",
-                **_urbansim_output_facet_meta(
-                    USIM_DATASTORE_H5, forecast_year=state.forecast_year
-                ),
-            )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="urbansim",
-        phase="postprocess",
-        outputs_class=UrbanSimPostprocessOutputs,
-        component_getter=lambda factory, state: factory.get_postprocessor(
-            "urbansim", state, WorkflowState.Stage.land_use
-        ),
-        component_executor=_execute_urbansim_postprocess,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "urbansim_postprocess", outputs
-        ),
-        output_logger=_log_outputs,
-    )
-
-
-def make_atlas_preprocess_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the ATLAS preprocess step function.
-
-    This step extracts UrbanSim HDF5 tables into ATLAS input CSVs and optionally
-    computes accessibility metrics required by the ATLAS vehicle ownership model.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing preprocess outputs.
-
-    Returns
-    -------
-    callable
-        Step function for ATLAS preprocess.
-    """
-    def _log_outputs(
-        outputs: AtlasPreprocessOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        run_scenario = getattr(getattr(settings, "atlas", None), "scenario", None)
-        _log_step_records(
-            record_items=outputs._iter_record_items(),
-            log_fn=log_output_only,
-            profile_schema_suffixes=(".csv", ".parquet"),
-            extra_meta_fn=lambda key, _path, _description: _atlas_artifact_facet_meta(
-                key,
-                run_scenario=run_scenario,
-                forecast_year=state.forecast_year,
-                artifact_family="atlas_preprocess_output",
-            ),
-        )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="atlas",
-        phase="preprocess",
-        outputs_class=AtlasPreprocessOutputs,
-        component_getter=lambda factory, state: factory.get_preprocessor(
-            "atlas", state, WorkflowState.Stage.vehicle_ownership_model
-        ),
-        component_executor=_execute_preprocess,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "atlas_preprocess", outputs
-        ),
-        output_logger=_log_outputs,
-    )
-
-
-def make_atlas_run_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the ATLAS run step function.
-
-    This step runs ATLAS to simulate vehicle ownership for the sub-year and
-    produces household/vehicle output CSVs.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing run outputs.
-
-    Returns
-    -------
-    callable
-        Step function for ATLAS run.
-    """
-    def _log_inputs(
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> Dict[str, Any]:
-        upstream = holder.atlas_preprocess
-        if upstream is None:
-            raise RuntimeError("ATLAS preprocess must complete first")
-        run_scenario = getattr(getattr(settings, "atlas", None), "scenario", None)
-        _log_step_records(
-            record_items=upstream._iter_record_items(),
-            log_fn=log_input_only,
-            profile_schema_suffixes=(".csv", ".parquet"),
-            extra_meta_fn=lambda key, _path, _description: _atlas_artifact_facet_meta(
-                key,
-                run_scenario=run_scenario,
-                forecast_year=state.forecast_year,
-                artifact_family="atlas_run_input",
-            ),
-        )
-        return {}
-
-    def _log_outputs(
-        outputs: AtlasRunOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        _log_step_records(
-            record_items=outputs._iter_record_items(),
-            log_fn=log_output_only,
-            profile_schema_suffixes=(".csv", ".parquet"),
-        )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="atlas",
-        phase="run",
-        outputs_class=AtlasRunOutputs,
-        component_getter=lambda factory, state: factory.get_runner(
-            "atlas", state, WorkflowState.Stage.vehicle_ownership_model
-        ),
-        component_executor=_execute_atlas_run,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "atlas_run", outputs
-        ),
-        input_logger=_log_inputs,
-        output_logger=_log_outputs,
-    )
-
-
-def make_atlas_postprocess_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the ATLAS postprocess step function.
-
-    This step updates the UrbanSim HDF5 datastore with ATLAS vehicle ownership
-    results and writes vehicles2 outputs used by BEAM.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing postprocess outputs.
-
-    Returns
-    -------
-    callable
-        Step function for ATLAS postprocess.
-    """
-
-    def _log_outputs(
-        outputs: AtlasPostprocessOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        _log_step_records(
-            record_items=outputs._iter_record_items(),
-            log_fn=log_output_only,
-            profile_schema_suffixes=(".csv", ".parquet"),
-        )
-        if outputs.usim_datastore_h5 is not None:
-            log_and_set_output(
-                key=USIM_DATASTORE_H5,
-                path=str(outputs.usim_datastore_h5),
-                description=(
-                    "UrbanSim datastore updated by ATLAS for year "
-                    f"{state.forecast_year}"
-                ),
-                coupler=coupler,
-                profile_file_schema=True,
-                h5_container=True,
-                hash_tables="if_unchanged",
-                **_urbansim_output_facet_meta(
-                    USIM_DATASTORE_H5, forecast_year=state.forecast_year
-                ),
-            )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="atlas",
-        phase="postprocess",
-        outputs_class=AtlasPostprocessOutputs,
-        component_getter=lambda factory, state: factory.get_postprocessor(
-            "atlas", state, WorkflowState.Stage.vehicle_ownership_model
-        ),
-        component_executor=_execute_atlas_postprocess,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "atlas_postprocess", outputs
-        ),
-        output_logger=_log_outputs,
-    )
-
-
-def make_activitysim_compile_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the ActivitySim compile step function.
-
-    This step compiles ActivitySim inputs (prepared by preprocess) into
-    optimized skim artifacts (Zarr) used by ActivitySim runs and BEAM.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-
-    Returns
-    -------
-    callable
-        Step function for ActivitySim compile.
-    """
-
-    @require_common_runtime("expected_outputs")
-    def _run_activitysim_compile_step(
-        *,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        expected_outputs: Dict[str, Any],
-    ) -> None:
-        factory = ModelFactory()
-        from workflow_state import WorkflowState as _WorkflowState
-
-        compile_runner = factory.get_runner(
-            "activitysim_compile",
-            state,
-            major_stage=_WorkflowState.Stage.activity_demand,
-        )
-
-        upstream = outputs_holder.activitysim_preprocess
-        if upstream is None:
-            raise RuntimeError(
-                "ActivitySim compile must run after activitysim_preprocess"
-            )
-        input_store = upstream.to_record_store()
-        omx_record = None
-        if input_store:
-            for record in input_store.all_records():
-                if getattr(record, "short_name", None) == ASIM_OMX_SKIMS:
-                    omx_record = record
-                    break
-        if omx_record is not None:
-            input_store = RecordStore(recordList=[omx_record])
-        else:
-            input_store = RecordStore()
-        if omx_record is not None:
-            omx_path = omx_record.get_absolute_path(base_path=workspace.full_path)
-            if omx_path and os.path.exists(omx_path):
-                cr.log_input(
-                    omx_path,
-                    key=ASIM_OMX_SKIMS,
-                    description="ActivitySim compile input skims (OMX)",
-                )
-        compile_outputs = compile_runner.run(input_store, workspace)
-
-        zarr_record = None
-        if compile_outputs:
-            for record in compile_outputs.all_records():
-                if record.short_name == ZARR_SKIMS:
-                    zarr_record = record
-                    break
-        zarr_output_path = expected_outputs.get(ZARR_SKIMS)
-        if not zarr_output_path and zarr_record is not None:
-            zarr_output_path = zarr_record.file_path
-        if zarr_output_path and os.path.exists(zarr_output_path):
-            log_and_set_output(
-                key=ZARR_SKIMS,
-                path=zarr_output_path,
-                description="ActivitySim compiled zarr skims",
-                coupler=coupler,
-                **_activitysim_output_facet_meta(
-                    ZARR_SKIMS,
-                    year=state.forecast_year,
-                    iteration=state.iteration,
-                ),
-            )
-
-    return _decorate_step_with_consist(
-        step_func=_run_activitysim_compile_step,
-        step_model="activitysim_compile",
-        description="activitysim compile workflow step",
-        outputs=[ZARR_SKIMS],
-        schema_outputs=[ZARR_SKIMS],
-        tags=["activitysim", "compile"],
-    )
-
-
-def make_activitysim_preprocess_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the ActivitySim preprocess step function.
-
-    This step prepares ActivitySim inputs from UrbanSim outputs and ensures
-    the mutable input directory contains the required tables and skims.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing preprocess outputs.
-
-    Returns
-    -------
-    callable
-        Step function for ActivitySim preprocess.
-
-    Notes
-    -----
-    This step focuses on preparing ActivitySim inputs. Config canonicalization
-    is also performed here so config ingestion is tied to the preprocess phase.
-    """
-
-    def _log_inputs(
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> Dict[str, Any]:
-        """
-        Log ActivitySim preprocess inputs.
-
-        This helper logs the UrbanSim datastore input if it exists in the coupler.
-
-        Parameters
-        ----------
-        settings : PilatesConfig
-            Simulation settings for config root resolution.
-        state : WorkflowState
-            Current workflow state (used for log metadata only).
-        workspace : Workspace
-            Workspace used to resolve mutable config paths.
-        holder : StepOutputsHolder
-            Outputs holder (unused for this helper).
-
-        Returns
-        -------
-        dict
-            Extra runtime kwargs for the step executor (empty for this helper).
-        """
-        usim_input = None
-        get_value = getattr(coupler, "get", None)
-        if callable(get_value):
-            updated_value = get_value(USIM_H5_UPDATED)
-            if updated_value is not None:
-                selected_key = USIM_H5_UPDATED
-                selected_value = updated_value
-            else:
-                current_value = get_value(USIM_DATASTORE_CURRENT_H5)
-                if current_value is not None:
-                    selected_key = USIM_DATASTORE_CURRENT_H5
-                    selected_value = current_value
-                else:
-                    selected_key = USIM_DATASTORE_BASE_H5
-                    selected_value = get_value(USIM_DATASTORE_BASE_H5)
-            usim_input = resolve_artifact_from_value(
-                selected_value,
-                key=selected_key,
-                workspace=workspace,
-            )
-        usim_path = artifact_to_path(usim_input, workspace)
-        if usim_path and os.path.exists(usim_path):
-            input_key = (
-                USIM_H5_UPDATED
-                if callable(get_value) and get_value(USIM_H5_UPDATED) is not None
-                else (
-                    USIM_DATASTORE_CURRENT_H5
-                    if callable(get_value)
-                    and get_value(USIM_DATASTORE_CURRENT_H5) is not None
-                    else USIM_DATASTORE_BASE_H5
-                )
-            )
-            input_desc = (
-                f"UrbanSim datastore updated by ATLAS for ActivitySim year {state.year}"
-                if input_key == USIM_H5_UPDATED
-                else (
-                    f"UrbanSim current datastore for ActivitySim year {state.year}"
-                    if input_key == USIM_DATASTORE_CURRENT_H5
-                    else f"UrbanSim base datastore for ActivitySim year {state.year}"
-                )
-            )
-            h5_tables_used = [
-                "households",
-                "persons",
-                "jobs",
-                "blocks",
-            ]
-            start_year = state.start_year
-            if start_year is not None:
-                h5_tables_used.extend(
-                    [
-                        f"/{start_year}/households",
-                        f"/{start_year}/persons",
-                        f"/{start_year}/jobs",
-                        f"/{start_year}/blocks",
-                    ]
-                )
-            log_and_set_input(
-                key=input_key,
-                path=usim_path,
-                description=input_desc,
-                coupler=coupler,
-                profile_file_schema=True,
-                h5_container=True,
-                h5_tables_used=h5_tables_used,
-            )
-        return {}
-
-    def _log_outputs(
-        outputs: ActivitySimPreprocessOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        """
-        Log ActivitySim preprocess outputs and update the coupler.
-
-        Parameters
-        ----------
-        outputs : ActivitySimPreprocessOutputs
-            Typed outputs containing ActivitySim input tables and skims.
-        settings : PilatesConfig
-            Simulation settings (unused for output logging).
-        state : WorkflowState
-            Current workflow state (used for log metadata only).
-        workspace : Workspace
-            Workspace for path resolution.
-        holder : StepOutputsHolder
-            Outputs holder (unused for this helper).
-        """
-        _log_step_records(
-            record_items=outputs._iter_record_items(),
-            log_fn=lambda key, path, description, **meta: log_and_set_output(
-                key=key,
-                path=path,
-                description=description,
-                coupler=coupler,
-                **meta,
-            ),
-            profile_schema_keys={ASIM_HOUSEHOLDS_IN, ASIM_PERSONS_IN, ASIM_LAND_USE_IN},
-        )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="activitysim",
-        phase="preprocess",
-        outputs_class=ActivitySimPreprocessOutputs,
-        component_getter=lambda factory, state: factory.get_preprocessor(
-            "activitysim", state, WorkflowState.Stage.activity_demand
-        ),
-        component_executor=_execute_preprocess,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "activitysim_preprocess", outputs
-        ),
-        input_logger=_log_inputs,
-        output_logger=_log_outputs,
-    )
-
-
-def make_activitysim_run_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the ActivitySim run step function.
-
-    This step executes the activity demand model for the year/iteration and
-    produces household/person/tour outputs consumed by BEAM and postprocessing.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing run outputs.
-
-    Returns
-    -------
-    callable
-        Step function for ActivitySim run.
-    """
-
-    def _log_inputs(
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> Dict[str, Any]:
-        upstream = holder.activitysim_preprocess
-        if upstream is None:
-            raise RuntimeError("ActivitySim preprocess must complete first")
-
-        profile_keys = {ASIM_HOUSEHOLDS_IN, ASIM_PERSONS_IN, ASIM_LAND_USE_IN}
-        for short_name, path, description in upstream._iter_record_items():
-            if short_name == ASIM_OMX_SKIMS:
-                continue
-            meta: Dict[str, Any] = {}
-            if short_name in profile_keys:
-                meta["profile_file_schema"] = True
-            log_and_set_input(
-                key=short_name,
-                path=str(path),
-                description=description,
-                coupler=coupler,
-                **meta,
-            )
-
-        extra_inputs = RecordStore()
-        zarr_value = None
-        get_value = getattr(coupler, "get", None)
-        if callable(get_value):
-            zarr_value = resolve_artifact_from_value(
-                get_value(ZARR_SKIMS),
-                key=ZARR_SKIMS,
-                workspace=workspace,
-            )
-        zarr_path = artifact_to_path(zarr_value, workspace)
-        if not zarr_path:
-            candidate = os.path.join(
-                workspace.get_asim_output_dir(), "cache", "skims.zarr"
-            )
-            if os.path.exists(candidate):
-                zarr_path = candidate
-        if zarr_path and os.path.exists(zarr_path):
-            extra_inputs.add_record(
-                FileRecord(
-                    file_path=zarr_path,
-                    short_name=ZARR_SKIMS,
-                    description="Compiled ActivitySim skims (Zarr)",
-                )
-            )
-            log_and_set_input(
-                key=ZARR_SKIMS,
-                path=zarr_path,
-                description=(
-                    f"ActivitySim compiled skims for year {state.year}, iter {state.iteration}"
-                ),
-                coupler=coupler,
-            )
-        return {"extra_inputs": extra_inputs}
-
-    def _log_outputs(
-        outputs: ActivitySimRunOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        for short_name, path, description in outputs._iter_record_items():
-            artifact = cr.log_output(
-                str(path),
-                key=short_name,
-                description=description,
-                **_activitysim_output_facet_meta(
-                    short_name,
-                    year=state.forecast_year,
-                    iteration=state.iteration,
-                ),
-            )
-            if artifact is not None and getattr(artifact, "hash", None):
-                outputs.raw_output_hashes[short_name] = artifact.hash
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="activitysim",
-        phase="run",
-        outputs_class=ActivitySimRunOutputs,
-        component_getter=lambda factory, state: factory.get_runner(
-            "activitysim", state, WorkflowState.Stage.activity_demand
-        ),
-        component_executor=_execute_run,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "activitysim_run", outputs
-        ),
-        input_logger=_log_inputs,
-        output_logger=_log_outputs,
-    )
-
-
-def make_activitysim_postprocess_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the ActivitySim postprocess step function.
-
-    This step converts ActivitySim outputs into downstream inputs and updates
-    the UrbanSim datastore for the next model stages.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing postprocess outputs.
-
-    Returns
-    -------
-    callable
-        Step function for ActivitySim postprocess.
-    """
-
-    def _log_outputs(
-        outputs: ActivitySimPostprocessOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        def _extra_meta(short_name: str, _path: str, _description: str) -> Dict[str, Any]:
-            meta: Dict[str, Any] = _activitysim_output_facet_meta(
-                short_name,
-                year=state.forecast_year,
-                iteration=state.iteration,
-            )
-            content_hash = outputs.processed_output_hashes.get(short_name)
-            if content_hash:
-                meta["content_hash"] = content_hash
-            return meta
-
-        _log_step_records(
-            record_items=outputs._iter_record_items(),
-            log_fn=log_output_only,
-            profile_schema_keys={
-                "persons_asim_out",
-                "trips_asim_out",
-                "tours_asim_out",
-                "beam_plans_asim_out",
-                "households_asim_out",
-            },
-            extra_meta_fn=_extra_meta,
-        )
-        if outputs.usim_datastore_h5 is not None:
-            log_and_set_output(
-                key=USIM_DATASTORE_H5,
-                path=str(outputs.usim_datastore_h5),
-                description=(
-                    f"UrbanSim datastore updated by ActivitySim for year {state.forecast_year}"
-                ),
-                coupler=coupler,
-                profile_file_schema=True,
-                h5_container=True,
-                hash_tables="if_unchanged",
-            )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="activitysim",
-        phase="postprocess",
-        outputs_class=ActivitySimPostprocessOutputs,
-        component_getter=lambda factory, state: factory.get_postprocessor(
-            "activitysim", state, WorkflowState.Stage.activity_demand
-        ),
-        component_executor=_execute_postprocess,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "activitysim_postprocess", outputs
-        ),
-        output_logger=_log_outputs,
-    )
-
-
-def make_beam_preprocess_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the BEAM preprocess step function.
-
-    This step builds the BEAM scenario inputs by transforming ActivitySim
-    demand outputs, adding ATLAS vehicles (if enabled), and staging warm-start
-    artifacts such as linkstats.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing preprocess outputs.
-
-    Returns
-    -------
-    callable
-        Step function for BEAM preprocess.
-
-    Notes
-    -----
-    This step focuses on generating BEAM inputs and canonicalizing BEAM config.
-    """
-
-    def _log_outputs(
-        outputs: BeamPreprocessOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        """
-        Log BEAM preprocess outputs and update the coupler.
-
-        This helper logs prepared BEAM input artifacts into the coupler for
-        downstream BEAM run and postprocess steps.
-
-        Parameters
-        ----------
-        outputs : BeamPreprocessOutputs
-            Typed outputs containing prepared BEAM inputs.
-        settings : PilatesConfig
-            Simulation settings for config root resolution.
-        state : WorkflowState
-            Current workflow state (used for log metadata only).
-        workspace : Workspace
-            Workspace used to resolve mutable BEAM config paths.
-        holder : StepOutputsHolder
-            Outputs holder (unused for this helper).
-        """
-        _log_step_records(
-            record_items=(
-                (
-                    key,
-                    path,
-                    f"BEAM prepared input {key} for year {state.year}, iter {state.iteration}",
-                )
-                for key, path in outputs.prepared_inputs.items()
-            ),
-            log_fn=lambda key, path, description, **meta: log_and_set_output(
-                key=key,
-                path=path,
-                description=description,
-                coupler=coupler,
-                **meta,
-            ),
-            profile_schema_keys={
-                "households_beam_in",
-                "persons_beam_in",
-                "plans_beam_in",
-            },
-        )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="beam",
-        phase="preprocess",
-        outputs_class=BeamPreprocessOutputs,
-        component_getter=lambda factory, state: factory.get_preprocessor(
-            "beam", state, WorkflowState.Stage.traffic_assignment
-        ),
-        component_executor=_execute_beam_preprocess,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "beam_preprocess", outputs
-        ),
-        output_logger=_log_outputs,
-    )
-
-
-def make_beam_run_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the BEAM run step function.
-
-    This step performs the traffic assignment simulation for the current
-    iteration and produces linkstats, skims, plans, and event outputs.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing run outputs.
-
-    Returns
-    -------
-    callable
-        Step function for BEAM run.
-    """
-
-    def _log_inputs(
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> Dict[str, Any]:
-        upstream = holder.beam_preprocess
-        if upstream is None:
-            raise RuntimeError("BEAM preprocess must complete first")
-
-        tracker = cr.current_tracker()
-        if tracker is not None:
-            _log_beam_r5_osm_input(
-                tracker=tracker,
-                settings=settings,
-                workspace=workspace,
-            )
-        for short_name, path, description in upstream._iter_record_items():
-            log_and_set_input(
-                key=short_name,
-                path=str(path),
-                description=description,
-                coupler=coupler,
-            )
-
-        from pathlib import Path
-
-        output_root = Path(workspace.get_beam_output_dir()) / settings.run.region
-        plans_path, experienced_path = find_last_run_output_plans(
-            output_root, "year-"
-        )
-        if plans_path is not None and plans_path.exists():
-            if plans_path.name == "output_plans.xml.gz":
-                plans_key = BEAM_OUTPUT_PLANS_XML
-            else:
-                plans_key = BEAM_PLANS_OUT
-            log_and_set_input(
-                key=plans_key,
-                path=str(plans_path),
-                description=(
-                    "BEAM warm-start plans (selected by BEAM from previous outputs)"
-                ),
-                coupler=coupler,
-            )
-        if experienced_path is not None and experienced_path.exists():
-            if experienced_path.name == "output_experienced_plans.xml.gz":
-                experienced_key = BEAM_OUTPUT_EXPERIENCED_PLANS_XML
-            else:
-                experienced_key = BEAM_EXPERIENCED_PLANS_XML
-            log_and_set_input(
-                key=experienced_key,
-                path=str(experienced_path),
-                description=(
-                    "BEAM warm-start experienced plans (selected by BEAM from previous outputs)"
-                ),
-                coupler=coupler,
-            )
-        return {}
-
-    def _log_outputs(
-        outputs: BeamRunOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        def _beam_run_extra_meta(
-            short_name: str,
-            _path: str,
-            _description: str,
-        ) -> Dict[str, Any]:
-            meta: Dict[str, Any] = {}
-            facet_meta = _beam_log_facet_meta(short_name)
-            if facet_meta:
-                meta.update(facet_meta)
-            if short_name.startswith("beam_network_final"):
-                meta.update(
-                    {
-                        "profile_file_schema": "if_changed",
-                        "reuse_if_unchanged": True,
-                        "reuse_scope": "any_uri",
-                    }
-                )
-                try:
-                    from pilates.database.schema.beam_schema import BeamNetworkFinal
-                except Exception:
-                    BeamNetworkFinal = None
-                if BeamNetworkFinal is not None:
-                    meta["schema"] = BeamNetworkFinal
-            return meta
-
-        _log_step_records(
-            record_items=outputs._iter_record_items(),
-            log_fn=log_output_only,
-            extra_meta_fn=_beam_run_extra_meta,
-        )
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="beam",
-        phase="run",
-        outputs_class=BeamRunOutputs,
-        component_getter=lambda factory, state: factory.get_runner(
-            "beam", state, WorkflowState.Stage.traffic_assignment
-        ),
-        component_executor=_execute_beam_run,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "beam_run", outputs
-        ),
-        input_logger=_log_inputs,
-        output_logger=_log_outputs,
-    )
-
-
-def make_beam_postprocess_step(
-    *,
-    coupler: CouplerProtocol,
-    outputs_holder: StepOutputsHolder,
-) -> Callable[..., None]:
-    """
-    Build the BEAM postprocess step function.
-
-    This step merges BEAM outputs into updated skims and produces final
-    skim artifacts for ActivitySim and UrbanSim inputs.
-
-    Parameters
-    ----------
-    coupler : object
-        Consist coupler for input/output logging.
-    outputs_holder : StepOutputsHolder
-        Holder for storing postprocess outputs.
-
-    Returns
-    -------
-    callable
-        Step function for BEAM postprocess.
-    """
-
-    def _log_outputs(
-        outputs: BeamPostprocessOutputs,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-        holder: StepOutputsHolder,
-    ) -> None:
-        for short_name, path, description in outputs._iter_record_items():
-            log_and_set_output(
-                key=short_name,
-                path=str(path),
-                description=description,
-                coupler=coupler,
-            )
-        for short_name, path in outputs.split_events.items():
-            facet_meta = _beam_postprocess_split_facet_meta(short_name)
-            log_output_only(
-                key=short_name,
-                path=str(path),
-                description=f"BEAM events parquet split ({short_name})",
-                profile_file_schema=True,
-                **facet_meta,
-            )
-        for short_name, path in outputs.split_event_links.items():
-            facet_meta = _beam_postprocess_split_facet_meta(short_name)
-            log_output_only(
-                key=short_name,
-                path=str(path),
-                description=f"BEAM events link table ({short_name})",
-                profile_file_schema=True,
-                **facet_meta,
-            )
-        upstream = holder.beam_run
-        if upstream is None:
-            return
-        combined_outputs = RecordStore()
-        combined_outputs += upstream.to_record_store()
-        combined_outputs += outputs.to_record_store()
-        update_coupler_from_beam_outputs(combined_outputs, coupler, workspace)
-
-    return _make_generic_step_function(
-        coupler=coupler,
-        outputs_holder=outputs_holder,
-        model_name="beam",
-        phase="postprocess",
-        outputs_class=BeamPostprocessOutputs,
-        component_getter=lambda factory, state: factory.get_postprocessor(
-            "beam", state, WorkflowState.Stage.traffic_assignment
-        ),
-        component_executor=_execute_beam_postprocess,
-        outputs_holder_setter=lambda holder, outputs: setattr(
-            holder, "beam_postprocess", outputs
-        ),
-        output_logger=_log_outputs,
-    )
-
-
-def make_postprocessing_step() -> Callable[..., None]:
-    """
-    Build the postprocessing step function.
-
-    This step runs optional post-run cleanup/export routines such as event
-    processing and copying outputs to external destinations.
-
-    Returns
-    -------
-    callable
-        Step function for postprocessing.
-    """
-
-    @require_common_runtime()
-    def _run_postprocessing_step(
-        *,
-        settings: PilatesConfig,
-        state: WorkflowState,
-        workspace: Workspace,
-    ) -> None:
-        if "postprocessing" in settings:
-            from pilates.postprocessing.postprocessor import (
-                copy_outputs_to_mep,
-                process_event_file,
-            )
-
-            process_event_file(settings, state.forecast_year, state.current_inner_iter)
-            copy_outputs_to_mep(
-                settings,
-                state.forecast_year,
-                state.current_inner_iter,
-                workspace,
-            )
-
-    return _decorate_step_with_consist(
-        step_func=_run_postprocessing_step,
-        step_model="postprocessing",
-        description="postprocessing workflow step",
-        tags=["postprocessing"],
-    )
