+++ tests/test_activitysim_postprocess_logging.py
@@ -2,6 +2,7 @@ from types import SimpleNamespace
 
 from pilates.activitysim.outputs import ActivitySimPostprocessOutputs
 from pilates.workflows import steps
+from pilates.workflows.steps import activitysim as steps_activitysim
 
 
 def test_activitysim_postprocess_logs_content_hash(monkeypatch, tmp_path) -> None:
@@ -12,7 +13,9 @@ def test_activitysim_postprocess_logs_content_hash(monkeypatch, tmp_path) -> Non
         return lambda *args, **inner_kwargs: None
 
     monkeypatch.setattr(
-        steps, "_make_generic_step_function", _fake_make_generic_step_function
+        steps_activitysim,
+        "_make_generic_step_function",
+        _fake_make_generic_step_function,
     )
 
     steps.make_activitysim_postprocess_step(
@@ -26,7 +29,7 @@ def test_activitysim_postprocess_logs_content_hash(monkeypatch, tmp_path) -> Non
     def _log_output_only(*, key, path, description, **meta):
         calls.append((key, meta))
 
-    monkeypatch.setattr(steps, "log_output_only", _log_output_only)
+    monkeypatch.setattr(steps_activitysim, "log_output_only", _log_output_only)
 
     outputs = ActivitySimPostprocessOutputs(
         usim_datastore_h5=None,
@@ -42,9 +45,11 @@ def test_activitysim_postprocess_logs_content_hash(monkeypatch, tmp_path) -> Non
     output_logger(
         outputs,
         settings=SimpleNamespace(),
-        state=SimpleNamespace(),
+        state=SimpleNamespace(forecast_year=0, iteration=0),
         workspace=SimpleNamespace(),
         holder=SimpleNamespace(),
     )
 
-    assert calls == [("asim_input_skims_zarr_archived", {"content_hash": "abc123"})]
+    assert len(calls) == 1
+    assert calls[0][0] == "asim_input_skims_zarr_archived"
+    assert calls[0][1]["content_hash"] == "abc123"
